+++ src/Actions/GetInvoiceStatusAction.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Xve\LaravelPeppol\Actions;
 
+use Xve\LaravelPeppol\Events\InvoiceStatusRetrieved;
 use Xve\LaravelPeppol\Exceptions\AuthenticationException;
 use Xve\LaravelPeppol\Exceptions\ConnectionException;
 use Xve\LaravelPeppol\Exceptions\InvoiceException;
@@ -25,7 +26,11 @@ class GetInvoiceStatusAction
                 throw InvoiceException::notFound($id);
             }
 
-            return InvoiceStatus::fromResponse($response->json());
+            $status = InvoiceStatus::fromResponse($response->json());
+
+            InvoiceStatusRetrieved::dispatch($id, $status);
+
+            return $status;
         } catch (\Illuminate\Http\Client\ConnectionException $e) {
             throw ConnectionException::unreachable();
         }